Trying to rename Gui to gui in 2 commits
Or else Windows WILL break stuff
