+++ src/main/java/nl/yacht/lakesideresort/controller/asdf/Command.java
@@ -1,4 +1,4 @@
-package nl.yacht.lakesideresort.controller.Gui;
+package nl.yacht.lakesideresort.controller.asdf;
 
 import java.io.IOException;
 import java.lang.reflect.Method;
